Feature ETP-273: Minor fixes
